refactor(v010): carve voicegw export-costs out of _legacy
Eleventh per-command carve-out in section 2.

Adds voicegateway/cli/export_costs.py (170 LOC) owning the export
schema and the export-costs command surface: _EXPORT_COLUMNS,
_EXPORT_KEY_MAP, _format_export_value, _format_export_row, and
export_costs_cmd.

The shared _parse_iso_date_arg helper stays in _legacy.py for now:
both export-costs and reconcile call it. Once reconcile is also
carved (next iteration) the helper moves to a dedicated _helpers
module in the section-2 finalisation commit.

Implementation note: the Python slice that removed the export-costs
region used _EXPORT_COLUMNS as the start marker and the next
@app.command (reconcile) as the stop marker. _parse_iso_date_arg
sits between them in line order and got swallowed; restored in the
same commit. Documented in the journal.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package (12 source
files now).
